scraper/requirements.txt

## c) `scraper/requirements.txt`

```txt
requests==2.32.3
pandas==2.2.2
SQLAlchemy==2.0.36
python-dateutil==2.9.0.post0
